Refactoring and trying to connect to livenet using docker
