+++ src/server/models/uplinks.model.ts
@@ -1,0 +1,8 @@
+export interface UplinkOptions
+{
+    testnet: boolean;
+    secret: string;
+    connectorPort: number;
+    adminApiPort: number;
+    allowedOrigins: Array<string>;
+}
